Return Home Modified
The return home functions were modified because the robot is no longer
constantly updating it's location. It now drives the calculated distance
watching for a wall on the front instead of constantly comparing it's
location to the desired location. The driving to candle function was
also modified to update its location after a certain period of time
using the angle it was driving at, and then re-scanning and turning if
necessary to a new angle.