Sonar Fix
Find Flame works - NOTE: the value at which the sensor reacts to seeing
the flame was changed to 800. Use to be 300.
Sonar Fix: pin 19 was broken, so bypassed it by operating the side sonar
with polling using pin 25 for the Sonar Output. Appears to run as well
as it did on an interrupt.